feat: 영화 예외처리 추가 (MovieNotFoundException)

+++ backend/src/main/java/com/mymovie/backend/movie/MovieService.java
@@ -21,15 +21,21 @@ public class MovieService {
     }
 
     public Movie updateMovie(Long id, Movie updated) {
-        Movie movie = movieRepository.findById(id).orElseThrow();
+        Movie movie = movieRepository.findById(id).orElseThrow(() -> new MovieNotFoundException(id));
         movie.setTitle(updated.getTitle());
         movie.setRating(updated.getRating());
         movie.setMemo(updated.getMemo());
         movie.setWatchedDate(updated.getWatchedDate());
+        movie.setDate(updated.getPoster());
+        movie.setPoster(updated.getPoster());
+        movie.setGenres(updated.getGenres());
         return movieRepository.save(movie);
     }
 
     public void deleteMovie(Long id) {
+        if (!movieRepository.existsById(id)) {
+            throw new MovieNotFoundException(id);
+        }
         movieRepository.deleteById(id);
     }
 }
